Update branding assets and database config
Replaced the favicon and sidebar logo with new assets in the frontend, updating references in index.html and Sidebar.jsx. Changed the displayed user name in Header.jsx to 'Belleza Plena Spa'. Updated the MySQL password in backend/settings.py from 'spabellezaplena' to 'root'. Replace password 'root' for 'spabellezaplena' in the other divice.